Drop subscription-manager and Extras repo enablement for RHEL
This broke my use of k-a because due to the way Ansible works, even if
the hosts are already subscribed, it's required to create the
`rhn_username` files on each host.

Everyone using RHEL should understand how to set up subscriptions.
(If they don't, we've failed)

That said there might be a case for sanity checking that the Extras
repo is enabled...but on the other hand things will quickly error out
due to missing packages, and that's hopefully obvious.